# Faster merging procedure, more detailed metrics analysis and more compatible for MOT16/17