--- a/motmetrics/apps/evaluateTracking.py
+++ b/motmetrics/apps/evaluateTracking.py
@@ -60,9 +60,10 @@ string in the seqmap.""", formatter_class=argparse.RawTextHelpFormatter)
     parser.add_argument('--fmt', type=str, help='Data format', default='mot15-2D')
     parser.add_argument('--solver', type=str, help='LAP solver to use')
     parser.add_argument('--skip', type=int, default=0, help='skip frames n means choosing one frame for every (n+1) frames')
+    parser.add_argument('--iou', type=float, default=0.5, help='special IoU threshold requirement for small targets')
     return parser.parse_args()
 
-def compare_dataframes(gts, ts, vsflag = ''):
+def compare_dataframes(gts, ts, vsflag = '', iou = 0.5):
     accs = []
     anas = []
     names = []
@@ -73,7 +74,7 @@ def compare_dataframes(gts, ts, vsflag = ''):
                 fd = open(vsflag+'/'+k+'.log','w')
             else:
                 fd = ''
-            acc, ana = mm.utils.CLEAR_MOT_M(gts[k][0], tsacc, gts[k][1], 'iou', distth=0.5, vflag=fd)
+            acc, ana = mm.utils.CLEAR_MOT_M(gts[k][0], tsacc, gts[k][1], 'iou', distth=iou, vflag=fd)
             if fd!='':
                 fd.close()
             accs.append(acc)
@@ -153,7 +154,7 @@ if __name__ == '__main__':
 
     mh = mm.metrics.create()
     st = time.time()
-    accs, analysis, names = compare_dataframes(gt, ts, args.log)
+    accs, analysis, names = compare_dataframes(gt, ts, args.log, 1.-args.iou)
     logging.info('adding frames: %.3f seconds.'%(time.time()-st))
     
     logging.info('Running metrics')