- origin:
````sh
06:36:18 INFO - Found 11 groundtruths and 11 test files.
06:36:18 INFO - Available LAP solvers ['scipy']
06:36:18 INFO - Default LAP solver 'scipy'
06:36:18 INFO - Loading files.
06:36:18 INFO - Comparing Venice-2...
06:36:20 INFO - Comparing ETH-Pedcross2...
06:36:21 INFO - Comparing TUD-Campus...
06:36:21 INFO - Comparing ADL-Rundle-6...
06:36:22 INFO - Comparing KITTI-13...
06:36:23 INFO - Comparing PETS09-S2L1...
06:36:24 INFO - Comparing TUD-Stadtmitte...
06:36:25 INFO - Comparing ETH-Bahnhof...
06:36:26 INFO - Comparing ADL-Rundle-8...
06:36:28 INFO - Comparing ETH-Sunnyday...
06:36:29 INFO - Comparing KITTI-17...
06:36:29 INFO - Running metrics
                IDF1   IDP   IDR  Rcll  Prcn  GT MT  PT  ML   FP    FN IDs   FM  MOTA  MOTP
Venice-2       35.5% 43.6% 29.9% 42.0% 61.3%  26  4  16   6 1890  4144  42   52 14.9% 0.274
ETH-Pedcross2  16.9% 61.3%  9.8% 12.2% 76.0% 133  1  17 115  240  5502  17   21  8.0% 0.301
TUD-Campus     55.8% 73.0% 45.1% 58.2% 94.1%   8  1   6   1   13   150   7    7 52.6% 0.277
ADL-Rundle-6   36.3% 50.0% 28.5% 42.1% 73.9%  24  1  18   5  744  2898  53   45 26.2% 0.275
KITTI-13       29.6% 38.8% 23.9% 28.1% 45.6%  42  0  17  25  255   548   4    5 -5.9% 0.318
PETS09-S2L1    66.6% 66.6% 66.6% 91.8% 91.8%  19 18   1   0  368   367  31   37 82.9% 0.269
TUD-Stadtmitte 64.5% 82.0% 53.1% 60.9% 94.0%  10  5   4   1   45   452   7    6 56.4% 0.346
ETH-Bahnhof    54.2% 60.2% 49.3% 58.9% 72.0% 171 41  55  75 1241  2223  17   69 35.7% 0.282
ADL-Rundle-8   28.3% 33.9% 24.3% 40.1% 55.8%  28  5  13  10 2152  4064  37   45  7.8% 0.261
ETH-Sunnyday   49.9% 83.2% 35.7% 37.1% 86.4%  30  7   4  19  108  1169   1    4 31.2% 0.240
KITTI-17       55.8% 72.2% 45.5% 52.4% 83.1%   9  1   7   1   73   325   6   10 40.8% 0.277
OVERALL        41.2% 53.2% 33.6% 45.3% 71.7% 500 84 158 258 7129 21842 222  301 26.8% 0.276
06:39:41 INFO - Completed
````
- this implement
```sh
06:38:00 INFO - Found 11 groundtruths and 11 test files.
06:38:00 INFO - Available LAP solvers ['scipy']
06:38:00 INFO - Default LAP solver 'scipy'
06:38:00 INFO - Loading files.
06:38:01 INFO - Comparing Venice-2...
06:38:02 INFO - Comparing ETH-Pedcross2...
06:38:03 INFO - Comparing TUD-Campus...
06:38:04 INFO - Comparing ADL-Rundle-6...
06:38:05 INFO - Comparing KITTI-13...
06:38:05 INFO - Comparing PETS09-S2L1...
06:38:07 INFO - Comparing TUD-Stadtmitte...
06:38:07 INFO - Comparing ETH-Bahnhof...
06:38:09 INFO - Comparing ADL-Rundle-8...
06:38:11 INFO - Comparing ETH-Sunnyday...
06:38:11 INFO - Comparing KITTI-17...
06:38:11 INFO - Running metrics
06:38:21 INFO - partials: 9.172 seconds.
06:38:21 INFO - mergeOverall: 9.173 seconds.
                IDF1   IDP   IDR  Rcll  Prcn  GT MT  PT  ML   FP    FN IDs   FM  MOTA  MOTP IDt IDa IDm
Venice-2       35.5% 43.6% 29.9% 42.0% 61.3%  26  4  16   6 1890  4144  42   52 14.9% 0.274  12  31   1
ETH-Pedcross2  16.9% 61.3%  9.8% 12.2% 76.0% 133  1  17 115  240  5502  17   21  8.0% 0.301  10  14   7
TUD-Campus     55.8% 73.0% 45.1% 58.2% 94.1%   8  1   6   1   13   150   7    7 52.6% 0.277   2   7   2
ADL-Rundle-6   36.3% 50.0% 28.5% 42.1% 73.9%  24  1  18   5  744  2898  53   45 26.2% 0.275  13  46   6
KITTI-13       29.6% 38.8% 23.9% 28.1% 45.6%  42  0  17  25  255   548   4    5 -5.9% 0.318   8   2   6
PETS09-S2L1    66.6% 66.6% 66.6% 91.8% 91.8%  19 18   1   0  368   367  31   37 82.9% 0.269   8  24   1
TUD-Stadtmitte 64.5% 82.0% 53.1% 60.9% 94.0%  10  5   4   1   45   452   7    6 56.4% 0.346   5   4   2
ETH-Bahnhof    54.2% 60.2% 49.3% 58.9% 72.0% 171 41  55  75 1241  2223  17   69 35.7% 0.282  37  14  34
ADL-Rundle-8   28.3% 33.9% 24.3% 40.1% 55.8%  28  5  13  10 2152  4064  37   45  7.8% 0.261   2  37   2
ETH-Sunnyday   49.9% 83.2% 35.7% 37.1% 86.4%  30  7   4  19  108  1169   1    4 31.2% 0.240   1   1   1
KITTI-17       55.8% 72.2% 45.5% 52.4% 83.1%   9  1   7   1   73   325   6   10 40.8% 0.277   1   6   1
OVERALL        41.2% 53.2% 33.6% 45.3% 71.7% 500 84 158 258 7129 21842 222  301 26.8% 0.276  99 186  63
06:38:21 INFO - Completed
```
### Faster merging procedure for results
- 9+ times faster when evaluating. 21 secs this version and 192 secs for the origin one.

### More detailed metric analysis
- Provide ID_transfer, ID_ascend, ID_migrate for ID mistakes analysis. 

### More compatible for MOT16/17
- Add new app in running MOT16/17 results, run motmetrics.apps.evaluateTracking