@Helicopt 2019-04-17
> @Helicopt merged, thanks! I wonder if you could update the readme to reflect the new app you added?

Sure. But it will take some time.